change: devide components for chash flow and profit loss
