+++ application/front/src/components/cashFlowBarChart/props.ts
@@ -1,0 +1,6 @@
+export interface CashFlowBarChartProps {
+  openingBalance: number;
+  operatingActivitiesCashFlow: number;
+  investingActivitiesCashFlow: number;
+  financingActivitiesCashFlow: number;
+}
