+++ application/front/src/components/profitLossBarChart/props.ts
@@ -1,0 +1,5 @@
+export interface ProfitLossBarChartProps {
+  sales: number;
+  costOfSales: number;
+  sellingAdministrativeExpense: number;
+}
