+++ application/front/src/components/waterFlowBarChart/props.ts
@@ -1,0 +1,11 @@
+export interface WaterFlowBarChartElement {
+  name: string;
+  value: number;
+  previousSum: number;
+}
+
+export interface WaterFlowBarChartProps {
+  data: WaterFlowBarChartElement[];
+  positiveColor: string;
+  negativeColor: string;
+}
