+++ application/front/src/constants/tooltip.ts
@@ -1,0 +1,5 @@
+export const tooltipStyle = {
+  backgroundColor: '#F6F4EB',
+  opacity: '0.8',
+  padding: '10px',
+};
